requirements.txt
django_storages
boto3
dj-database-url
openpyxl